perf(ci): shard Temporal workflow integration tests
