+++ .github/workflows/ci.yml
@@ -73,7 +73,7 @@ jobs:
               '[range(0; $total) | {index: ., number: (. + 1), total: $total}]'
           }
           echo "unit_matrix=$(matrix "$(jq '.unitTests | length' impact-plan.json)" 4)" >> "$GITHUB_OUTPUT"
-          echo "integration_matrix=$(matrix "$(jq '.integrationTests | length' impact-plan.json)" 3)" >> "$GITHUB_OUTPUT"
+          echo "integration_matrix=$(matrix "$(jq '.integrationTests | length' impact-plan.json)" 6)" >> "$GITHUB_OUTPUT"
           echo "e2e_matrix=$(matrix "$(jq '.e2eTests | length' impact-plan.json)" 3)" >> "$GITHUB_OUTPUT"
           {
             echo '### Change-impact plan'
