+++ package.json
@@ -35,7 +35,7 @@
     "format:check": "oxfmt --check",
     "test": "bun test",
     "test:unit": "bun test",
-    "test:integration": "bun test --timeout 30000 ./test/integration/api.integration.ts && bun test --timeout 60000 ./test/integration/file-upload.integration.ts && bun test --timeout 30000 ./test/integration/db.integration.ts && bun test --timeout 30000 ./test/integration/nats.integration.ts && bun test --timeout 30000 ./test/integration/selfhosted-control-transport.integration.ts && bun test --timeout 120000 ./test/integration/selfhosted-auth-callout.integration.ts && bun test --timeout 30000 ./test/integration/workspace-isolation.integration.ts && bun test --timeout 30000 ./test/integration/temporal-workflow.integration.ts && bun test --timeout 30000 ./test/integration/worker-activity.integration.ts && bun test --timeout 30000 ./test/integration/worker-restart.integration.ts",
+    "test:integration": "bun test --timeout 30000 ./test/integration/api.integration.ts && bun test --timeout 60000 ./test/integration/file-upload.integration.ts && bun test --timeout 30000 ./test/integration/db.integration.ts && bun test --timeout 30000 ./test/integration/nats.integration.ts && bun test --timeout 30000 ./test/integration/selfhosted-control-transport.integration.ts && bun test --timeout 120000 ./test/integration/selfhosted-auth-callout.integration.ts && bun test --timeout 30000 ./test/integration/workspace-isolation.integration.ts && bun test --timeout 30000 ./test/integration/temporal-workflow-activities.integration.ts && bun test --timeout 30000 ./test/integration/temporal-workflow-capacity.integration.ts && bun test --timeout 120000 ./test/integration/temporal-workflow-continuity.integration.ts && bun test --timeout 30000 ./test/integration/temporal-workflow-dispatch.integration.ts && bun test --timeout 30000 ./test/integration/temporal-workflow-interrupt.integration.ts && bun test --timeout 180000 ./test/integration/temporal-workflow-worker-death.integration.ts && bun test --timeout 30000 ./test/integration/worker-activity.integration.ts && bun test --timeout 30000 ./test/integration/worker-restart.integration.ts",
     "test:e2e": "bun scripts/run-browser-e2e.ts && bun test ./test/e2e/sandbox.e2e.ts && bun test --timeout 360000 ./test/e2e/rig-setup.e2e.ts && bun test --timeout 360000 ./test/e2e/rig-verification.e2e.ts && bun test --timeout 360000 ./test/e2e/channel-a.e2e.ts && bun test --timeout 2100000 ./apps/worker/test/desktop-image.e2e.ts",
     "test:live": "bun test ./test/live/*.live.ts",
     "test:all": "bun run test:unit && bun run test:integration && bun run test:e2e",
